Consolidate glean.ts into one sync file
